requirements.txt
anthropic==0.18.0
pdf-parse==1.1.1
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
jsonschema==4.20.0
click==8.1.7
pathlib2==2.3.7 